move the router part of the compile_test to new abi_router_test.py, and add fixtures for teal outputs
