+++ pyteal/compiler/compiler_test.py
@@ -2193,2434 +2193,3 @@ retsub
 
     with pytest.raises(pt.TealInternalError):
         pt.compileTeal(program_access_b4_store_broken, pt.Mode.Application, version=6)
-
-
-def test_router_app():
-    def add_methods_to_router(router: pt.Router):
-        @pt.ABIReturnSubroutine
-        def add(
-            a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64
-        ) -> pt.Expr:
-            return output.set(a.get() + b.get())
-
-        meth = router.add_method_handler(add)
-        assert meth.method_signature() == "add(uint64,uint64)uint64"
-
-        @pt.ABIReturnSubroutine
-        def sub(
-            a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64
-        ) -> pt.Expr:
-            return output.set(a.get() - b.get())
-
-        meth = router.add_method_handler(sub)
-        assert meth.method_signature() == "sub(uint64,uint64)uint64"
-
-        @pt.ABIReturnSubroutine
-        def mul(
-            a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64
-        ) -> pt.Expr:
-            return output.set(a.get() * b.get())
-
-        meth = router.add_method_handler(mul)
-        assert meth.method_signature() == "mul(uint64,uint64)uint64"
-
-        @pt.ABIReturnSubroutine
-        def div(
-            a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64
-        ) -> pt.Expr:
-            return output.set(a.get() / b.get())
-
-        meth = router.add_method_handler(div)
-        assert meth.method_signature() == "div(uint64,uint64)uint64"
-
-        @pt.ABIReturnSubroutine
-        def mod(
-            a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64
-        ) -> pt.Expr:
-            return output.set(a.get() % b.get())
-
-        meth = router.add_method_handler(mod)
-        assert meth.method_signature() == "mod(uint64,uint64)uint64"
-
-        @pt.ABIReturnSubroutine
-        def all_laid_to_args(
-            _a: pt.abi.Uint64,
-            _b: pt.abi.Uint64,
-            _c: pt.abi.Uint64,
-            _d: pt.abi.Uint64,
-            _e: pt.abi.Uint64,
-            _f: pt.abi.Uint64,
-            _g: pt.abi.Uint64,
-            _h: pt.abi.Uint64,
-            _i: pt.abi.Uint64,
-            _j: pt.abi.Uint64,
-            _k: pt.abi.Uint64,
-            _l: pt.abi.Uint64,
-            _m: pt.abi.Uint64,
-            _n: pt.abi.Uint64,
-            _o: pt.abi.Uint64,
-            _p: pt.abi.Uint64,
-            *,
-            output: pt.abi.Uint64,
-        ):
-            return output.set(
-                _a.get()
-                + _b.get()
-                + _c.get()
-                + _d.get()
-                + _e.get()
-                + _f.get()
-                + _g.get()
-                + _h.get()
-                + _i.get()
-                + _j.get()
-                + _k.get()
-                + _l.get()
-                + _m.get()
-                + _n.get()
-                + _o.get()
-                + _p.get()
-            )
-
-        meth = router.add_method_handler(all_laid_to_args)
-        assert (
-            meth.method_signature()
-            == "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
-        )
-
-        @pt.ABIReturnSubroutine
-        def empty_return_subroutine() -> pt.Expr:
-            return pt.Log(pt.Bytes("appear in both approval and clear state"))
-
-        meth = router.add_method_handler(
-            empty_return_subroutine,
-            method_config=pt.MethodConfig(
-                no_op=pt.CallConfig.CALL,
-                opt_in=pt.CallConfig.ALL,
-                clear_state=pt.CallConfig.CALL,
-            ),
-        )
-        assert meth.method_signature() == "empty_return_subroutine()void"
-
-        @pt.ABIReturnSubroutine
-        def log_1(*, output: pt.abi.Uint64) -> pt.Expr:
-            return output.set(1)
-
-        meth = router.add_method_handler(
-            log_1,
-            method_config=pt.MethodConfig(
-                no_op=pt.CallConfig.CALL,
-                opt_in=pt.CallConfig.CALL,
-                clear_state=pt.CallConfig.CALL,
-            ),
-        )
-
-        assert meth.method_signature() == "log_1()uint64"
-
-        @pt.ABIReturnSubroutine
-        def log_creation(*, output: pt.abi.String) -> pt.Expr:
-            return output.set("logging creation")
-
-        meth = router.add_method_handler(
-            log_creation, method_config=pt.MethodConfig(no_op=pt.CallConfig.CREATE)
-        )
-        assert meth.method_signature() == "log_creation()string"
-
-        @pt.ABIReturnSubroutine
-        def approve_if_odd(condition_encoding: pt.abi.Uint32) -> pt.Expr:
-            return (
-                pt.If(condition_encoding.get() % pt.Int(2))
-                .Then(pt.Approve())
-                .Else(pt.Reject())
-            )
-
-        meth = router.add_method_handler(
-            approve_if_odd,
-            method_config=pt.MethodConfig(
-                no_op=pt.CallConfig.NEVER, clear_state=pt.CallConfig.CALL
-            ),
-        )
-        assert meth.method_signature() == "approve_if_odd(uint32)void"
-
-    on_completion_actions = pt.BareCallActions(
-        opt_in=pt.OnCompleteAction.call_only(pt.Log(pt.Bytes("optin call"))),
-        clear_state=pt.OnCompleteAction.call_only(pt.Approve()),
-    )
-
-    with pytest.raises(pt.TealInputError) as e:
-        pt.Router("will-error", on_completion_actions).compile_program(
-            version=6, optimize=pt.OptimizeOptions(frame_pointers=True)
-        )
-
-    assert "Frame pointers aren't available" in str(e.value)
-
-    _router_with_oc = pt.Router(
-        "ASimpleQuestionablyRobustContract", on_completion_actions
-    )
-    add_methods_to_router(_router_with_oc)
-    (
-        actual_ap_with_oc_compiled,
-        actual_csp_with_oc_compiled,
-        _,
-    ) = _router_with_oc.compile_program(version=6)
-
-    expected_ap_with_oc = """#pragma version 6
-txn NumAppArgs
-int 0
-==
-bnz main_l20
-txna ApplicationArgs 0
-method "add(uint64,uint64)uint64"
-==
-bnz main_l19
-txna ApplicationArgs 0
-method "sub(uint64,uint64)uint64"
-==
-bnz main_l18
-txna ApplicationArgs 0
-method "mul(uint64,uint64)uint64"
-==
-bnz main_l17
-txna ApplicationArgs 0
-method "div(uint64,uint64)uint64"
-==
-bnz main_l16
-txna ApplicationArgs 0
-method "mod(uint64,uint64)uint64"
-==
-bnz main_l15
-txna ApplicationArgs 0
-method "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
-==
-bnz main_l14
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l13
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l12
-txna ApplicationArgs 0
-method "log_creation()string"
-==
-bnz main_l11
-err
-main_l11:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-==
-&&
-assert
-callsub logcreation_8
-store 67
-byte 0x151f7c75
-load 67
-concat
-log
-int 1
-return
-main_l12:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-txn ApplicationID
-int 0
-!=
-&&
-||
-assert
-callsub log1_7
-store 65
-byte 0x151f7c75
-load 65
-itob
-concat
-log
-int 1
-return
-main_l13:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-||
-assert
-callsub emptyreturnsubroutine_6
-int 1
-return
-main_l14:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 30
-txna ApplicationArgs 2
-btoi
-store 31
-txna ApplicationArgs 3
-btoi
-store 32
-txna ApplicationArgs 4
-btoi
-store 33
-txna ApplicationArgs 5
-btoi
-store 34
-txna ApplicationArgs 6
-btoi
-store 35
-txna ApplicationArgs 7
-btoi
-store 36
-txna ApplicationArgs 8
-btoi
-store 37
-txna ApplicationArgs 9
-btoi
-store 38
-txna ApplicationArgs 10
-btoi
-store 39
-txna ApplicationArgs 11
-btoi
-store 40
-txna ApplicationArgs 12
-btoi
-store 41
-txna ApplicationArgs 13
-btoi
-store 42
-txna ApplicationArgs 14
-btoi
-store 43
-txna ApplicationArgs 15
-store 46
-load 46
-int 0
-extract_uint64
-store 44
-load 46
-int 8
-extract_uint64
-store 45
-load 30
-load 31
-load 32
-load 33
-load 34
-load 35
-load 36
-load 37
-load 38
-load 39
-load 40
-load 41
-load 42
-load 43
-load 44
-load 45
-callsub alllaidtoargs_5
-store 47
-byte 0x151f7c75
-load 47
-itob
-concat
-log
-int 1
-return
-main_l15:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 24
-txna ApplicationArgs 2
-btoi
-store 25
-load 24
-load 25
-callsub mod_4
-store 26
-byte 0x151f7c75
-load 26
-itob
-concat
-log
-int 1
-return
-main_l16:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 18
-txna ApplicationArgs 2
-btoi
-store 19
-load 18
-load 19
-callsub div_3
-store 20
-byte 0x151f7c75
-load 20
-itob
-concat
-log
-int 1
-return
-main_l17:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 12
-txna ApplicationArgs 2
-btoi
-store 13
-load 12
-load 13
-callsub mul_2
-store 14
-byte 0x151f7c75
-load 14
-itob
-concat
-log
-int 1
-return
-main_l18:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 6
-txna ApplicationArgs 2
-btoi
-store 7
-load 6
-load 7
-callsub sub_1
-store 8
-byte 0x151f7c75
-load 8
-itob
-concat
-log
-int 1
-return
-main_l19:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 0
-txna ApplicationArgs 2
-btoi
-store 1
-load 0
-load 1
-callsub add_0
-store 2
-byte 0x151f7c75
-load 2
-itob
-concat
-log
-int 1
-return
-main_l20:
-txn OnCompletion
-int OptIn
-==
-bnz main_l22
-err
-main_l22:
-txn ApplicationID
-int 0
-!=
-assert
-byte "optin call"
-log
-int 1
-return
-
-// add
-add_0:
-store 4
-store 3
-load 3
-load 4
-+
-store 5
-load 5
-retsub
-
-// sub
-sub_1:
-store 10
-store 9
-load 9
-load 10
--
-store 11
-load 11
-retsub
-
-// mul
-mul_2:
-store 16
-store 15
-load 15
-load 16
-*
-store 17
-load 17
-retsub
-
-// div
-div_3:
-store 22
-store 21
-load 21
-load 22
-/
-store 23
-load 23
-retsub
-
-// mod
-mod_4:
-store 28
-store 27
-load 27
-load 28
-%
-store 29
-load 29
-retsub
-
-// all_laid_to_args
-alllaidtoargs_5:
-store 63
-store 62
-store 61
-store 60
-store 59
-store 58
-store 57
-store 56
-store 55
-store 54
-store 53
-store 52
-store 51
-store 50
-store 49
-store 48
-load 48
-load 49
-+
-load 50
-+
-load 51
-+
-load 52
-+
-load 53
-+
-load 54
-+
-load 55
-+
-load 56
-+
-load 57
-+
-load 58
-+
-load 59
-+
-load 60
-+
-load 61
-+
-load 62
-+
-load 63
-+
-store 64
-load 64
-retsub
-
-// empty_return_subroutine
-emptyreturnsubroutine_6:
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_7:
-int 1
-store 66
-load 66
-retsub
-
-// log_creation
-logcreation_8:
-byte 0x00106c6f6767696e67206372656174696f6e
-store 68
-load 68
-retsub""".strip()
-
-    assert expected_ap_with_oc == actual_ap_with_oc_compiled
-
-    expected_csp_with_oc = """#pragma version 6
-txn NumAppArgs
-int 0
-==
-bnz main_l8
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l7
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l6
-txna ApplicationArgs 0
-method "approve_if_odd(uint32)void"
-==
-bnz main_l5
-err
-main_l5:
-txna ApplicationArgs 1
-int 0
-extract_uint32
-store 2
-load 2
-callsub approveifodd_2
-int 1
-return
-main_l6:
-callsub log1_1
-store 1
-byte 0x151f7c75
-load 1
-itob
-concat
-log
-int 1
-return
-main_l7:
-callsub emptyreturnsubroutine_0
-int 1
-return
-main_l8:
-int 1
-return
-
-// empty_return_subroutine
-emptyreturnsubroutine_0:
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_1:
-int 1
-store 0
-load 0
-retsub
-
-// approve_if_odd
-approveifodd_2:
-store 3
-load 3
-int 2
-%
-bnz approveifodd_2_l2
-int 0
-return
-approveifodd_2_l2:
-int 1
-return""".strip()
-    assert expected_csp_with_oc == actual_csp_with_oc_compiled
-
-    _router_without_oc = pt.Router("yetAnotherContractConstructedFromRouter")
-    add_methods_to_router(_router_without_oc)
-    (
-        actual_ap_without_oc_compiled,
-        actual_csp_without_oc_compiled,
-        _,
-    ) = _router_without_oc.compile_program(version=6)
-    expected_ap_without_oc = """#pragma version 6
-txna ApplicationArgs 0
-method "add(uint64,uint64)uint64"
-==
-bnz main_l18
-txna ApplicationArgs 0
-method "sub(uint64,uint64)uint64"
-==
-bnz main_l17
-txna ApplicationArgs 0
-method "mul(uint64,uint64)uint64"
-==
-bnz main_l16
-txna ApplicationArgs 0
-method "div(uint64,uint64)uint64"
-==
-bnz main_l15
-txna ApplicationArgs 0
-method "mod(uint64,uint64)uint64"
-==
-bnz main_l14
-txna ApplicationArgs 0
-method "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
-==
-bnz main_l13
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l12
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l11
-txna ApplicationArgs 0
-method "log_creation()string"
-==
-bnz main_l10
-err
-main_l10:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-==
-&&
-assert
-callsub logcreation_8
-store 67
-byte 0x151f7c75
-load 67
-concat
-log
-int 1
-return
-main_l11:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-txn ApplicationID
-int 0
-!=
-&&
-||
-assert
-callsub log1_7
-store 65
-byte 0x151f7c75
-load 65
-itob
-concat
-log
-int 1
-return
-main_l12:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-||
-assert
-callsub emptyreturnsubroutine_6
-int 1
-return
-main_l13:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 30
-txna ApplicationArgs 2
-btoi
-store 31
-txna ApplicationArgs 3
-btoi
-store 32
-txna ApplicationArgs 4
-btoi
-store 33
-txna ApplicationArgs 5
-btoi
-store 34
-txna ApplicationArgs 6
-btoi
-store 35
-txna ApplicationArgs 7
-btoi
-store 36
-txna ApplicationArgs 8
-btoi
-store 37
-txna ApplicationArgs 9
-btoi
-store 38
-txna ApplicationArgs 10
-btoi
-store 39
-txna ApplicationArgs 11
-btoi
-store 40
-txna ApplicationArgs 12
-btoi
-store 41
-txna ApplicationArgs 13
-btoi
-store 42
-txna ApplicationArgs 14
-btoi
-store 43
-txna ApplicationArgs 15
-store 46
-load 46
-int 0
-extract_uint64
-store 44
-load 46
-int 8
-extract_uint64
-store 45
-load 30
-load 31
-load 32
-load 33
-load 34
-load 35
-load 36
-load 37
-load 38
-load 39
-load 40
-load 41
-load 42
-load 43
-load 44
-load 45
-callsub alllaidtoargs_5
-store 47
-byte 0x151f7c75
-load 47
-itob
-concat
-log
-int 1
-return
-main_l14:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 24
-txna ApplicationArgs 2
-btoi
-store 25
-load 24
-load 25
-callsub mod_4
-store 26
-byte 0x151f7c75
-load 26
-itob
-concat
-log
-int 1
-return
-main_l15:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 18
-txna ApplicationArgs 2
-btoi
-store 19
-load 18
-load 19
-callsub div_3
-store 20
-byte 0x151f7c75
-load 20
-itob
-concat
-log
-int 1
-return
-main_l16:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 12
-txna ApplicationArgs 2
-btoi
-store 13
-load 12
-load 13
-callsub mul_2
-store 14
-byte 0x151f7c75
-load 14
-itob
-concat
-log
-int 1
-return
-main_l17:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 6
-txna ApplicationArgs 2
-btoi
-store 7
-load 6
-load 7
-callsub sub_1
-store 8
-byte 0x151f7c75
-load 8
-itob
-concat
-log
-int 1
-return
-main_l18:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-txna ApplicationArgs 1
-btoi
-store 0
-txna ApplicationArgs 2
-btoi
-store 1
-load 0
-load 1
-callsub add_0
-store 2
-byte 0x151f7c75
-load 2
-itob
-concat
-log
-int 1
-return
-
-// add
-add_0:
-store 4
-store 3
-load 3
-load 4
-+
-store 5
-load 5
-retsub
-
-// sub
-sub_1:
-store 10
-store 9
-load 9
-load 10
--
-store 11
-load 11
-retsub
-
-// mul
-mul_2:
-store 16
-store 15
-load 15
-load 16
-*
-store 17
-load 17
-retsub
-
-// div
-div_3:
-store 22
-store 21
-load 21
-load 22
-/
-store 23
-load 23
-retsub
-
-// mod
-mod_4:
-store 28
-store 27
-load 27
-load 28
-%
-store 29
-load 29
-retsub
-
-// all_laid_to_args
-alllaidtoargs_5:
-store 63
-store 62
-store 61
-store 60
-store 59
-store 58
-store 57
-store 56
-store 55
-store 54
-store 53
-store 52
-store 51
-store 50
-store 49
-store 48
-load 48
-load 49
-+
-load 50
-+
-load 51
-+
-load 52
-+
-load 53
-+
-load 54
-+
-load 55
-+
-load 56
-+
-load 57
-+
-load 58
-+
-load 59
-+
-load 60
-+
-load 61
-+
-load 62
-+
-load 63
-+
-store 64
-load 64
-retsub
-
-// empty_return_subroutine
-emptyreturnsubroutine_6:
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_7:
-int 1
-store 66
-load 66
-retsub
-
-// log_creation
-logcreation_8:
-byte 0x00106c6f6767696e67206372656174696f6e
-store 68
-load 68
-retsub""".strip()
-    assert actual_ap_without_oc_compiled == expected_ap_without_oc
-
-    expected_csp_without_oc = """#pragma version 6
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l6
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l5
-txna ApplicationArgs 0
-method "approve_if_odd(uint32)void"
-==
-bnz main_l4
-err
-main_l4:
-txna ApplicationArgs 1
-int 0
-extract_uint32
-store 2
-load 2
-callsub approveifodd_2
-int 1
-return
-main_l5:
-callsub log1_1
-store 1
-byte 0x151f7c75
-load 1
-itob
-concat
-log
-int 1
-return
-main_l6:
-callsub emptyreturnsubroutine_0
-int 1
-return
-
-// empty_return_subroutine
-emptyreturnsubroutine_0:
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_1:
-int 1
-store 0
-load 0
-retsub
-
-// approve_if_odd
-approveifodd_2:
-store 3
-load 3
-int 2
-%
-bnz approveifodd_2_l2
-int 0
-return
-approveifodd_2_l2:
-int 1
-return""".strip()
-    assert actual_csp_without_oc_compiled == expected_csp_without_oc
-
-    _router_with_oc = pt.Router(
-        "QuestionableRouterGenerateCodeWithFramePointer", on_completion_actions
-    )
-    add_methods_to_router(_router_with_oc)
-    (
-        actual_ap_with_oc_compiled,
-        actual_csp_with_oc_compiled,
-        _,
-    ) = _router_with_oc.compile_program(version=8)
-
-    expected_ap_with_oc = """#pragma version 8
-txn NumAppArgs
-int 0
-==
-bnz main_l20
-txna ApplicationArgs 0
-method "add(uint64,uint64)uint64"
-==
-bnz main_l19
-txna ApplicationArgs 0
-method "sub(uint64,uint64)uint64"
-==
-bnz main_l18
-txna ApplicationArgs 0
-method "mul(uint64,uint64)uint64"
-==
-bnz main_l17
-txna ApplicationArgs 0
-method "div(uint64,uint64)uint64"
-==
-bnz main_l16
-txna ApplicationArgs 0
-method "mod(uint64,uint64)uint64"
-==
-bnz main_l15
-txna ApplicationArgs 0
-method "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
-==
-bnz main_l14
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l13
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l12
-txna ApplicationArgs 0
-method "log_creation()string"
-==
-bnz main_l11
-err
-main_l11:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-==
-&&
-assert
-callsub logcreationcaster_17
-int 1
-return
-main_l12:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-txn ApplicationID
-int 0
-!=
-&&
-||
-assert
-callsub log1caster_16
-int 1
-return
-main_l13:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-||
-assert
-callsub emptyreturnsubroutinecaster_15
-int 1
-return
-main_l14:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub alllaidtoargscaster_14
-int 1
-return
-main_l15:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub modcaster_13
-int 1
-return
-main_l16:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub divcaster_12
-int 1
-return
-main_l17:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub mulcaster_11
-int 1
-return
-main_l18:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub subcaster_10
-int 1
-return
-main_l19:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub addcaster_9
-int 1
-return
-main_l20:
-txn OnCompletion
-int OptIn
-==
-bnz main_l22
-err
-main_l22:
-txn ApplicationID
-int 0
-!=
-assert
-byte "optin call"
-log
-int 1
-return
-
-// add
-add_0:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-+
-frame_bury 0
-retsub
-
-// sub
-sub_1:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
--
-frame_bury 0
-retsub
-
-// mul
-mul_2:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-*
-frame_bury 0
-retsub
-
-// div
-div_3:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-/
-frame_bury 0
-retsub
-
-// mod
-mod_4:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-%
-frame_bury 0
-retsub
-
-// all_laid_to_args
-alllaidtoargs_5:
-proto 16 1
-int 0
-frame_dig -16
-frame_dig -15
-+
-frame_dig -14
-+
-frame_dig -13
-+
-frame_dig -12
-+
-frame_dig -11
-+
-frame_dig -10
-+
-frame_dig -9
-+
-frame_dig -8
-+
-frame_dig -7
-+
-frame_dig -6
-+
-frame_dig -5
-+
-frame_dig -4
-+
-frame_dig -3
-+
-frame_dig -2
-+
-frame_dig -1
-+
-frame_bury 0
-retsub
-
-// empty_return_subroutine
-emptyreturnsubroutine_6:
-proto 0 0
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_7:
-proto 0 1
-int 0
-int 1
-frame_bury 0
-retsub
-
-// log_creation
-logcreation_8:
-proto 0 1
-byte ""
-byte 0x00106c6f6767696e67206372656174696f6e
-frame_bury 0
-retsub
-
-// add_caster
-addcaster_9:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub add_0
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// sub_caster
-subcaster_10:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub sub_1
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// mul_caster
-mulcaster_11:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub mul_2
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// div_caster
-divcaster_12:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub div_3
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// mod_caster
-modcaster_13:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub mod_4
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// all_laid_to_args_caster
-alllaidtoargscaster_14:
-proto 0 0
-int 0
-dupn 16
-byte ""
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-txna ApplicationArgs 3
-btoi
-frame_bury 3
-txna ApplicationArgs 4
-btoi
-frame_bury 4
-txna ApplicationArgs 5
-btoi
-frame_bury 5
-txna ApplicationArgs 6
-btoi
-frame_bury 6
-txna ApplicationArgs 7
-btoi
-frame_bury 7
-txna ApplicationArgs 8
-btoi
-frame_bury 8
-txna ApplicationArgs 9
-btoi
-frame_bury 9
-txna ApplicationArgs 10
-btoi
-frame_bury 10
-txna ApplicationArgs 11
-btoi
-frame_bury 11
-txna ApplicationArgs 12
-btoi
-frame_bury 12
-txna ApplicationArgs 13
-btoi
-frame_bury 13
-txna ApplicationArgs 14
-btoi
-frame_bury 14
-txna ApplicationArgs 15
-frame_bury 17
-frame_dig 17
-int 0
-extract_uint64
-frame_bury 15
-frame_dig 17
-int 8
-extract_uint64
-frame_bury 16
-frame_dig 1
-frame_dig 2
-frame_dig 3
-frame_dig 4
-frame_dig 5
-frame_dig 6
-frame_dig 7
-frame_dig 8
-frame_dig 9
-frame_dig 10
-frame_dig 11
-frame_dig 12
-frame_dig 13
-frame_dig 14
-frame_dig 15
-frame_dig 16
-callsub alllaidtoargs_5
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// empty_return_subroutine_caster
-emptyreturnsubroutinecaster_15:
-proto 0 0
-callsub emptyreturnsubroutine_6
-retsub
-
-// log_1_caster
-log1caster_16:
-proto 0 0
-int 0
-callsub log1_7
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// log_creation_caster
-logcreationcaster_17:
-proto 0 0
-byte ""
-callsub logcreation_8
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-concat
-log
-retsub""".strip()
-    assert actual_ap_with_oc_compiled == expected_ap_with_oc
-
-    expected_csp_with_oc = """#pragma version 8
-txn NumAppArgs
-int 0
-==
-bnz main_l8
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l7
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l6
-txna ApplicationArgs 0
-method "approve_if_odd(uint32)void"
-==
-bnz main_l5
-err
-main_l5:
-callsub approveifoddcaster_5
-int 1
-return
-main_l6:
-callsub log1caster_4
-int 1
-return
-main_l7:
-callsub emptyreturnsubroutinecaster_3
-int 1
-return
-main_l8:
-int 1
-return
-
-// empty_return_subroutine
-emptyreturnsubroutine_0:
-proto 0 0
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_1:
-proto 0 1
-int 0
-int 1
-frame_bury 0
-retsub
-
-// approve_if_odd
-approveifodd_2:
-proto 1 0
-frame_dig -1
-int 2
-%
-bnz approveifodd_2_l2
-int 0
-return
-approveifodd_2_l2:
-int 1
-return
-
-// empty_return_subroutine_caster
-emptyreturnsubroutinecaster_3:
-proto 0 0
-callsub emptyreturnsubroutine_0
-retsub
-
-// log_1_caster
-log1caster_4:
-proto 0 0
-int 0
-callsub log1_1
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// approve_if_odd_caster
-approveifoddcaster_5:
-proto 0 0
-int 0
-txna ApplicationArgs 1
-int 0
-extract_uint32
-frame_bury 0
-frame_dig 0
-callsub approveifodd_2
-retsub""".strip()
-    assert actual_csp_with_oc_compiled == expected_csp_with_oc
-
-    _router_without_oc = pt.Router(
-        "yetAnotherContractConstructedFromRouterWithFramePointer"
-    )
-    add_methods_to_router(_router_without_oc)
-    (
-        actual_ap_without_oc_compiled,
-        actual_csp_without_oc_compiled,
-        _,
-    ) = _router_without_oc.compile_program(version=8)
-
-    expected_ap_without_oc = """#pragma version 8
-txna ApplicationArgs 0
-method "add(uint64,uint64)uint64"
-==
-bnz main_l18
-txna ApplicationArgs 0
-method "sub(uint64,uint64)uint64"
-==
-bnz main_l17
-txna ApplicationArgs 0
-method "mul(uint64,uint64)uint64"
-==
-bnz main_l16
-txna ApplicationArgs 0
-method "div(uint64,uint64)uint64"
-==
-bnz main_l15
-txna ApplicationArgs 0
-method "mod(uint64,uint64)uint64"
-==
-bnz main_l14
-txna ApplicationArgs 0
-method "all_laid_to_args(uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64,uint64)uint64"
-==
-bnz main_l13
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l12
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l11
-txna ApplicationArgs 0
-method "log_creation()string"
-==
-bnz main_l10
-err
-main_l10:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-==
-&&
-assert
-callsub logcreationcaster_17
-int 1
-return
-main_l11:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-txn ApplicationID
-int 0
-!=
-&&
-||
-assert
-callsub log1caster_16
-int 1
-return
-main_l12:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-txn OnCompletion
-int OptIn
-==
-||
-assert
-callsub emptyreturnsubroutinecaster_15
-int 1
-return
-main_l13:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub alllaidtoargscaster_14
-int 1
-return
-main_l14:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub modcaster_13
-int 1
-return
-main_l15:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub divcaster_12
-int 1
-return
-main_l16:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub mulcaster_11
-int 1
-return
-main_l17:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub subcaster_10
-int 1
-return
-main_l18:
-txn OnCompletion
-int NoOp
-==
-txn ApplicationID
-int 0
-!=
-&&
-assert
-callsub addcaster_9
-int 1
-return
-
-// add
-add_0:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-+
-frame_bury 0
-retsub
-
-// sub
-sub_1:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
--
-frame_bury 0
-retsub
-
-// mul
-mul_2:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-*
-frame_bury 0
-retsub
-
-// div
-div_3:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-/
-frame_bury 0
-retsub
-
-// mod
-mod_4:
-proto 2 1
-int 0
-frame_dig -2
-frame_dig -1
-%
-frame_bury 0
-retsub
-
-// all_laid_to_args
-alllaidtoargs_5:
-proto 16 1
-int 0
-frame_dig -16
-frame_dig -15
-+
-frame_dig -14
-+
-frame_dig -13
-+
-frame_dig -12
-+
-frame_dig -11
-+
-frame_dig -10
-+
-frame_dig -9
-+
-frame_dig -8
-+
-frame_dig -7
-+
-frame_dig -6
-+
-frame_dig -5
-+
-frame_dig -4
-+
-frame_dig -3
-+
-frame_dig -2
-+
-frame_dig -1
-+
-frame_bury 0
-retsub
-
-// empty_return_subroutine
-emptyreturnsubroutine_6:
-proto 0 0
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_7:
-proto 0 1
-int 0
-int 1
-frame_bury 0
-retsub
-
-// log_creation
-logcreation_8:
-proto 0 1
-byte ""
-byte 0x00106c6f6767696e67206372656174696f6e
-frame_bury 0
-retsub
-
-// add_caster
-addcaster_9:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub add_0
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// sub_caster
-subcaster_10:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub sub_1
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// mul_caster
-mulcaster_11:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub mul_2
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// div_caster
-divcaster_12:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub div_3
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// mod_caster
-modcaster_13:
-proto 0 0
-int 0
-dupn 2
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-frame_dig 1
-frame_dig 2
-callsub mod_4
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// all_laid_to_args_caster
-alllaidtoargscaster_14:
-proto 0 0
-int 0
-dupn 16
-byte ""
-txna ApplicationArgs 1
-btoi
-frame_bury 1
-txna ApplicationArgs 2
-btoi
-frame_bury 2
-txna ApplicationArgs 3
-btoi
-frame_bury 3
-txna ApplicationArgs 4
-btoi
-frame_bury 4
-txna ApplicationArgs 5
-btoi
-frame_bury 5
-txna ApplicationArgs 6
-btoi
-frame_bury 6
-txna ApplicationArgs 7
-btoi
-frame_bury 7
-txna ApplicationArgs 8
-btoi
-frame_bury 8
-txna ApplicationArgs 9
-btoi
-frame_bury 9
-txna ApplicationArgs 10
-btoi
-frame_bury 10
-txna ApplicationArgs 11
-btoi
-frame_bury 11
-txna ApplicationArgs 12
-btoi
-frame_bury 12
-txna ApplicationArgs 13
-btoi
-frame_bury 13
-txna ApplicationArgs 14
-btoi
-frame_bury 14
-txna ApplicationArgs 15
-frame_bury 17
-frame_dig 17
-int 0
-extract_uint64
-frame_bury 15
-frame_dig 17
-int 8
-extract_uint64
-frame_bury 16
-frame_dig 1
-frame_dig 2
-frame_dig 3
-frame_dig 4
-frame_dig 5
-frame_dig 6
-frame_dig 7
-frame_dig 8
-frame_dig 9
-frame_dig 10
-frame_dig 11
-frame_dig 12
-frame_dig 13
-frame_dig 14
-frame_dig 15
-frame_dig 16
-callsub alllaidtoargs_5
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// empty_return_subroutine_caster
-emptyreturnsubroutinecaster_15:
-proto 0 0
-callsub emptyreturnsubroutine_6
-retsub
-
-// log_1_caster
-log1caster_16:
-proto 0 0
-int 0
-callsub log1_7
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// log_creation_caster
-logcreationcaster_17:
-proto 0 0
-byte ""
-callsub logcreation_8
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-concat
-log
-retsub""".strip()
-    assert actual_ap_without_oc_compiled == expected_ap_without_oc
-
-    expected_csp_without_oc = """#pragma version 8
-txna ApplicationArgs 0
-method "empty_return_subroutine()void"
-==
-bnz main_l6
-txna ApplicationArgs 0
-method "log_1()uint64"
-==
-bnz main_l5
-txna ApplicationArgs 0
-method "approve_if_odd(uint32)void"
-==
-bnz main_l4
-err
-main_l4:
-callsub approveifoddcaster_5
-int 1
-return
-main_l5:
-callsub log1caster_4
-int 1
-return
-main_l6:
-callsub emptyreturnsubroutinecaster_3
-int 1
-return
-
-// empty_return_subroutine
-emptyreturnsubroutine_0:
-proto 0 0
-byte "appear in both approval and clear state"
-log
-retsub
-
-// log_1
-log1_1:
-proto 0 1
-int 0
-int 1
-frame_bury 0
-retsub
-
-// approve_if_odd
-approveifodd_2:
-proto 1 0
-frame_dig -1
-int 2
-%
-bnz approveifodd_2_l2
-int 0
-return
-approveifodd_2_l2:
-int 1
-return
-
-// empty_return_subroutine_caster
-emptyreturnsubroutinecaster_3:
-proto 0 0
-callsub emptyreturnsubroutine_0
-retsub
-
-// log_1_caster
-log1caster_4:
-proto 0 0
-int 0
-callsub log1_1
-frame_bury 0
-byte 0x151f7c75
-frame_dig 0
-itob
-concat
-log
-retsub
-
-// approve_if_odd_caster
-approveifoddcaster_5:
-proto 0 0
-int 0
-txna ApplicationArgs 1
-int 0
-extract_uint32
-frame_bury 0
-frame_dig 0
-callsub approveifodd_2
-retsub""".strip()
-
-    assert actual_csp_without_oc_compiled == expected_csp_without_oc
