+++ tests/integration/teal/router/questionable_clear_v6.teal
@@ -1,0 +1,70 @@
+#pragma version 6
+txn NumAppArgs
+int 0
+==
+bnz main_l8
+txna ApplicationArgs 0
+method "empty_return_subroutine()void"
+==
+bnz main_l7
+txna ApplicationArgs 0
+method "log_1()uint64"
+==
+bnz main_l6
+txna ApplicationArgs 0
+method "approve_if_odd(uint32)void"
+==
+bnz main_l5
+err
+main_l5:
+txna ApplicationArgs 1
+int 0
+extract_uint32
+store 2
+load 2
+callsub approveifodd_2
+int 1
+return
+main_l6:
+callsub log1_1
+store 1
+byte 0x151f7c75
+load 1
+itob
+concat
+log
+int 1
+return
+main_l7:
+callsub emptyreturnsubroutine_0
+int 1
+return
+main_l8:
+int 1
+return
+
+// empty_return_subroutine
+emptyreturnsubroutine_0:
+byte "appear in both approval and clear state"
+log
+retsub
+
+// log_1
+log1_1:
+int 1
+store 0
+load 0
+retsub
+
+// approve_if_odd
+approveifodd_2:
+store 3
+load 3
+int 2
+%
+bnz approveifodd_2_l2
+int 0
+return
+approveifodd_2_l2:
+int 1
+return
